postgres: Separate out common functions into own file
